Daily Update - Echo Strip progressing
The EFIR page is complete, I am currently working on the Mix page. I had to move the lights around otherwise the idea I have for the knobs wouldn't work.

The idea is to have the Dry vol and panning, and wet vol and panning on their own areas, with phase buttons attached to them dead centre. I'll have to use the smaller ADSR knobs to make it fit (the echo strip isn't as wide as I thought I could make it, but that's okay)

dummy feedback knobs are there for now.

After I'm done all this, the next step will be to decifer the steps tab. I have it pretty much figured out, and I have an idea of what I want to do.. I just have to figure out how Fermata does its scaling for that container. (I know the MD and Porta FM containers are smaller, I'll see if I can replicate it)